Unify OSRM backends for OSM and Mapbox plugins
This patch unifies the OSRM backend in both the OSM and Mapbox plugins,
adding some extra functionalities to QGeoRouteParserOsrmV5 to handle
the extra osrm-text-instructions information coming from the Mapbox
servers.
It also adds a plugin parameter to let the user choose whether to
use the server's text instructions or the plugin-generated ones.

Change-Id: Id7ce73f4285e2e7db6872f40d72c0610847fce91
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>