Apple platforms: properly guard API access when requesting location info
requestAlwaysAuthorization and requestWhenInUseAuthorization are
prohibited APIs on macOS, and the former is also prohibited API on tvOS.
This means they are akin to private APIs and must not be called under
any circumstances. The respondsToSelector calls are also removed as Qt
no longer supports OS versions where thee methods are not available,
and in any case such uses should be replaced with __builtin_available.

Background location updates are also now made available on watchOS 4 and
above.

Change-Id: I817d69f4ec71b96f03dda42635d675499e8908c0
Reviewed-by: Gabriel de Dietrich <gabriel.dedietrich@qt.io>