Fix fitViewportToMapItems in presence of MapQuickItems with zoomLevel
Map.fitViewportToMapItems currently uses the screen geometries of the
items to calculate the new values for zoom level and center.
When the zoomLevel property is set on MapQuickItems, their on-screen
geometry is affected by a hidden transformation within the item.
This patch considers that transformation in the calculations.

Task-number: QTBUG-63093
Change-Id: I328028f1901b31225866aa3d749849815d59a2a1
Reviewed-by: Vlad Seryakov <vseryakov@gmail.com>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>