Introduce QGeoMapObject
QGeoMapObject is a new class of map item, not deriving anymore from
QQuickItem, and designed to be more lightweight, and, most importantly,
to be easily backable by SDK-specific implementations, so to act as
an as thin as possible wrapper around those.

QGeoMapObject is intended to be the base class for this type of items.
This patch provides no mean to dynamically add/remove GeoMapObjects.
The intended way to do it is by using a MapObjectView, coming
initially with the Qt.labs.location extra plugin.

Change-Id: I8d6a45a4a32059c7ec4d904f75352e176bffda1e
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>