Introduce MapObjectView
This also adds addMapObject/removeMapObject to MapObjectView

With this patch, MapObjectView becomes the intended way to
programmatically populate a map with map objects. Objects can still
be declared inside a Map{} element, but to programmatically add them
it is necessary to add them to a MapObjectView that is added to the
map.
MapObjectView has then 3 ways to populate the map: via nested map
objects, via a model, and via programmatically added map objects.
All these 3 ways can be used simultaneously.

Change-Id: Ic7712355fa5e5787a0b83d9547693e288cba1960
Reviewed-by: BogDan Vatra <bogdan@kdab.com>