Fix Map.zoomLevel not honoring user-set values
Apparently m_userMinimumZoomLevel is set but not used when setting
the zoom level.
With this patch, the minimumZoomLevelChanged signal is still emitted
when the "implicit" minimumZoomLevel of the map changes (the one
derived from the map size), but only if no user minimum zoom level
is explicitly set.

Task-number: QTBUG-66107
Change-Id: Id7f87cbd237407b255deea90dad80eeadd2d2f0d
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>