Fix declarative_ui::MouseArea::test_enabled failing in CI
The test in the headline failed in CI on first runs, but succeeded in repeats
when it is run alone. This was because the previous test
declarative_ui::MouseArea::test_basic_press_release led to a situation where the
mouseClick() does not actually lead to a click event.

QTBUG-66534 was created to track that bug. Skip the part of the test that leads
to the failure for now.

Task-number: QTBUG-66533
Task-number: QTBUG-66534
Task-number: QTBUG-66216
Change-Id: Iadf3844e97603414594aceafcb1f36cf40653073
Reviewed-by: Paolo Angelelli <paolo.angelelli@qt.io>