API Fix: change QGeoRouteRequest::extraParameters into QVariantMap
Needed as it turns out the QGeoRouteRequest associated with a QGeoRoute
has to be exposed also through the declarative API (so
QDeclarativeGeoRoute must expose an associated
QDeclarativeGeoRouteRequest).
QDeclarativeGeoRouteRequest was, until now, intended to be only a
one-way API (to feed data into the plugin).
Now that it has also to expose data coming from the plugin, including
the associated extraParameters that do would come anymore from
nested MapParameters but from the QGeoRouteRequest itself.
For this reason the type has to be a QML-friently type.

Change-Id: I083e0fe97ad1088f6084665037003e24edb54371
Reviewed-by: Lars Knoll <lars.knoll@qt.io>