API Fix: Expose Waypoint metadata and RouteQuery extraParameters
Both RouteQuery and Waypoint are currently intended to be a one-way
interface to push data into a route request.
With the upcoming navigation support, there is need for these to become
also a way to report to the user updated request info.
This patch fixes the problem by adding two read only qml properties,
a Waypoint.metadata, that can also be used to read all of what is
being set via Map Parameters, and a RouteQuery.extraParameters,
intended to be used in the same way.

Change-Id: I65e484d96d50e89f44271781cb21e58dfe7badf8
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>