API Fix: expose the RouteQuery used to produce a Route
Currently, Route is intended to be a one-way interface
to push data into a route request.
With the upcoming navigation support, there is need for this to become
also a way to report to the user the updated request info,
for example when, Navigator.currentRoute changes.
This patch introduces a RouteQuery read-only qml property so it would
be possible to access the related query data in a route, and use it in
the UI.

Change-Id: Ibea8cb724313606feae614c61736ee14105d1045
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>