Allow threaded tile provisioning
This change virtualizes four key methods of QGeoTiledMappingManagerEngine,
that need to be reimplemented in order to be able to run tile provisioning
in a separate thread.
QGeoTileRequestManager is also now privately exported, as it has to be
used in those methods.
Finally the patch includes some minor cleanup in QGeoTileFetcher.

Change-Id: I3063263120f935f0cde54abf00bd1caa476ab215
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>