Fix crash in Map{Polyline,Polygon,Route}ObjectQsg
When these objects get added to a map but aren't immediately visible
(meaning are outside the visible region entirely), their respective
QSGNodes will be marked as blocked.
However, this isn't handled gracefully in qsgbatchrenderer.
Renderer::nodeWasAdded discards nodes that are blocked.
Meaning that when they will become unblocked, the node will
result in being unallocated.

This patch makes sure that a node is not instantiated at all
if it would result in a blocked node.

Change-Id: I8e174ba1d317e1d0c4006d993ec558bdad8a5b39
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>