Use QPointer to avoid dangling QDeclarativeGeoRoute in NavigatorPrivate
QDeclarativeNavigator::setRoute takes a QDeclarativeGeoRoute.
This type is not COW, and could be destroyed by the user or the route
model.

This patch stores it inside a QPointer.
It also stores the contained QGeoRoute, that is what the navigation
engines are supposed to use to perform navigation, and is COW.

Task-number: QTBUG-68536
Change-Id: I73fb79faeb3ac9efefcbd778ee106d29fbf26658
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>