Enable asynchronous incremental updates of QPlaceReply
This patch adds a new signal to QPlaceReply, contentUpdated.
This signal can be emitted in subclasses when the provider is capable
of notifying partial updates of the content.

The typical use case if fetching the data from multiple databases, also
locally instead of over network, and incrementally populating the
results, as opposed to the current way of operating, that is via previous
page / next page, in which case it's the backend who provides the proper
query parameters to obtain the previous or the next page.

This signal is currently handled only for QPlaceSearchReplies in
QDeclarativeSearchResultModel, where such a signal would trigger a
layout update.

Change-Id: I0aaba5aa9249b61c970d3a309d93fab7fc39d667
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>