Move management of QDeclarativeGeoMap camera data into QGeoMap
This patch makes QDeclarativeGeoMap always refer to QGeoMap
when it comes to setting and getting camera data properties.
The signal emission for each of these properties is also deferred
to the reception of QGeoMap::cameraDataChanged.
This patch also enable plugins to enforce tilt ranges per zoom,
updating the tilt value (as well as any other camera data value)
on their own.

Change-Id: Icc16645ea53fa2b53b33530f802ce390a2479d39
Reviewed-by: BogDan Vatra <bogdan@kdab.com>