Allow searching along a QGeoRoute
The result will be backend-dependent. If a backend does not support
this, it will be the same as not specifying a search area.
The current approach does not expose any new public API, but also
does not allow specifying the distance from the route.
If this will become necessary, public API might have to be
extended.

Task-number: QTBUG-69649
Change-Id: I9230d502a6cdb55c620e3a57a0e78cdb6f8955a6
Reviewed-by: Michael Brasser <michael.brasser@live.com>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>
Reviewed-by: BogDan Vatra <bogdan@kdab.com>