winrt: Check availability in requestAccess on desktop builds
For some reason the native status is always Disabled for desktop builds
(even though the location service is running).
So with desktop builds we have to use requestAccess to gain information
about the state of the location service. As the functionality is async the
important parts of QWinRTFunctions are shadowed locally.

Change-Id: I575ee0b161de734c29453a7a07350bc8d09720e8
Reviewed-by: Andre de la Rocha <andre.rocha@qt.io>
Reviewed-by: Miguel Costa <miguel.costa@qt.io>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>