winrt: Initialize position source on first usage
Loading the plugin without access to the location service will
work and not error out. It is still possible to enable location service
later and the backend will work as expected.

This patch changes the behavior so that the pop up asking for access to
the user's location data is no longer shown when the plugin is loaded, but
when the backend is actually used for the first time.

Change-Id: I23100f7867610c6f23b2d2ea5c15c268468949a9
Reviewed-by: Miguel Costa <miguel.costa@qt.io>