winrt: Remove incorrect asserts and fix incorrect condition
setPreferredPositioningMethods and setUpdateInterval are called by init so
these functions should not assert in initializing state.

Additionally the early exit condition in setPreferredPositioningMethods
has to be the same as in setUpdateInterval. Both cases have to return
early if the locator has not been initialized. This initialization
happens in init, so both states, Initializing and Initialized, are fine
for further handling.

Change-Id: I4ef8db7178c38299328041a8d163a1b9bbecc5b1
Reviewed-by: Andre de la Rocha <andre.rocha@qt.io>
Reviewed-by: Miguel Costa <miguel.costa@qt.io>