Expose QAbstractNavigator
This change exposes the QAbstractNavigator instantiated by
QDeclarativeNavigator. By doing so, it is possible to expose to the user
engine-specific features and data via properties, signals and methods
added to the QAbstractNavigator subclass.
However, while this approach greatly simplifies the exposure of
such features, it also leads to writing non-cross-plugin code.

Change-Id: I03be9b6d0868d940dc0ac80520f3b507bec2953a
Reviewed-by: Paolo Angelelli <paolo.angelelli@qt.io>