Move PluginParameter into PositioningQuick, expose also in QtPositioning
With this patch the QDeclarativeGeoServiceProviderParameter in the
location-private module becomes QDeclarativePluginParameter in the
positioningquick-private module.

This renamed class is then registered, under the same QML typename
"PluginParameter" both in QtPositioning and QtLocation qml plugins.

In this way it will become possible to use the same type both inside
QtLocation "Plugin" and inside QtPositioning "PositionSource",
and regardless of which is the QML plugin that provides
"PluginParameter", this will translate into the same cpp type.

[ChangeLog][QtPositioning] Exposed PluginParameter also from
QtPositioning qml plugin, to be used in PositionSource as replacement
of environment variables.

Task-number: QTBUG-66304
Change-Id: I1d982ff689130ae896c616480567aa12d9b49c3a
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>