QDeclarativeSearchResultModel: modify a map's value in-place instead of overwriting
More efficient, because the copy taken from the element in the
container must needs detach on the remove call. Calling remove on the
element in the container, OTOH, requires no detach (unless the page is
otherwise shared, of course).

The old code wouldn't detach the whole map when the int was out of
range, but that's not the normal case.

Change-Id: I2febd591ff78cea86949aa068938a3be4ca115f0
Reviewed-by: Paolo Angelelli <paolo.angelelli@qt.io>