QDeclarativeGeoMapParameter: replace the last use of QSignalMapper in the module
The code here is a bit tricky, as it relies on dynamic property
inspection, and therefore can't use the statically-typed Qt5
connection syntax to simply connect to a lambda to captures 'i' by
value.

But QSignalMapper is not just deprecated, its use here is also
overkill. Unlike your normal signal mapper, the code doesn't map
many source objects to different ints, there is actually one
QSignalMapper object _per sender_. But this use-case doesn't need
all the sender() inspection that QSignalMapper does under the hood.
It just needs a QObject with a map() slot that emits mapped(i),
where 'i' is a member variable.

Implement that.

Change-Id: Ib363144c8ab7997cb0a34c8c9c1a0c348fd38d57
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>