Fix flakiness of QGeoPositionInfoSource::createDefaultSource()
After ac899dcc4, createDefaultSource() uses the first available position
plugin to create an info source - even if that is not valid. The
non-deterministically sorted list of plugins results a flaky behavior,
especially when two plugins (geoclue2 and serialnmea) have the same priority
and one of them is unconfigured and not supposed to work.

Bring back the ability to find the first valid position plugin.

Change-Id: I82947e59252ac3727478fe14fd328338b54b4dea
Reviewed-by: Paolo Angelelli <paolo.angelelli@qt.io>