QtPositioning: adapt to Qt6
Make QtPositioning module buildable under Qt6:
- remove calls to deleted QMetaType-registration methods
- replace int -> qsizetype where it prevents the code from being
  compiled
- fix some deprecation warnings
- fix library dependencies in .pro file for building tests
- fix weatherinfo example

For now only qmake is supported. CMake support will be done in
scope of a separate patch.

Geoflickr example is yet to be ported, as it requires extracting
XmlListModel from xmlpatterns.

QtLocation is also skipped for now, as it seems to require
heavy refactoring due to changes in QSGMaterialShader interface.
QtLocation build is disabled in qmake for now.

Task-number: QTBUG-89853
Change-Id: I8cc4da6efabe32ce114701219d5431f8480bee3b
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>