Update commercial license headers
Updated header.COMM to the files in tqtc-qtlocation.
Examples, tests or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the line count causes failures in tests.

Task-number: QTQAINFRA-4167
Change-Id: Ie04ccedd48eacda57bfae6689dac78c1cd7f2ca1
Reviewed-by: Antti Kokko <antti.kokko@qt.io>