QtPositioning: refactor unit tests
Some of the QtPositioning QML unit tests can be separated from the
QtLocation tests, and run with QtPositioning and QtPositioningQuick
only.
This patch moves such unit tests into separate subdirectory.
It also updated the tests to reflect the changes in Qt6:
Since Qt6 the variant type behaves as var, so the tests that rely
on the obsolete behavior need to be updated.

The QGeoShape::center() method is no longer marked Q_INVOKABLE,
because with updated Qt6 behavior it overrides the center
properties of the child classes (like QGeoCircle or QGeoRectangle).
It's marked as a read-only property instread, so that it could
still be accessed from QML code.

[ChangeLog][QtPositioning][QGeoShape] center is now a read-only
property. The center() function is no longer a Q_INVOKABLE.

Task-number: QTBUG-90491
Task-number: QTBUG-90836
Change-Id: I3c7993aa8abe1199e825d46ab19ccfb53902018f
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>