QtPositioning: rename QGeoPolygon::setPath
Was planned to Qt6 some time back.
Renaming QGeoPolygon::setPath to setPerimiter, and also updating the
type of Q_PROPERTY to get rid of ambiguities.
Protected getter and setter for perimiter property are also removed,
as they are now not needed.

[ChangeLog][QtPositioning][QGeoPolygon] Rename QGeoPolygon::setPath()
and QGeoPolygon::path() to QGeoPolygon::setPerimiter() and
QGeoPolygon::perimiter() respectively.
Update the type of perimiter peoperty from QVariantList to
QList<QGeoCoordinate>.

Task-number: QTBUG-90491
Change-Id: I5f764c9fc7759c5bddd21a6d573c8af0e4ed4138
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>