NMEA: distinguish between different satellite systems
While working with device that supports multiple satellite systems, we
can receive GSV and GSA messages from different satellite system types.
Previously we did not distinguish the types, so our code was not able
to collect the full information.
For example:
1. We receive $GPGSV (GPS) and save them to the list of satellites in
   view.
2. We receive $GLGSV (GLONASS) - we overwrite the list instead of
   extending it.
And the situation was even more complicated with $GNGSA (satellites in
use from multiple systems) messages.

This patch tries to fix all these issues and introduce a proper
satellite system type handling.

Fixes: QTBUG-91074
Change-Id: I3fa03710c4318ebc2c10da1b676f7631751cabdb
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>