QtPositioning: increase control object lifetime in GYPSY plugin
For some reason the GObject control object was destroyed as soon as
the device was created.
With the current version of glib it results in immediate destruction
of all the created objects, so the newly created device got destroyed
as well. As a result, the plugin was unusable.
This patch extends the lifetime of the main control object for the
lifetime of the whole plugin.

Change-Id: I2ff2d40dea33b89c70f1f1188569f06fed4c88d9
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>