PositionSource: set preferred positioning methods to All by default
Preferred positioning methods in PositionSource are set to
NoPositioningMethods by default, which does not make much sense
for me, because it technically forbids receiving positioning data.
Normally this does not affect the applications, because the logic
for settings preferred methods handles NoPositioningMethods specially.
However on Android, if the application with PositionSource is launched
for the first time (and so needs to request permissions), the permissions
are not properly set, and so the positioning data is not received.
Application restart is required.

This patch sets default preferred positioning methods to
AllPositioningMethods, allowing android applications to work correctly.

Task-number: QTBUG-91528
Change-Id: I65b793e840dad3dd4f8acb1f3ce1a9267581489b
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>