QDeclarativePosition: add property bindings
Signals of QDeclarativePosition class are removed because they
were only needed for QML. With the new bindable properties we
do not need them any more.

[ChangeLog][QtPositioning][Position] horizontalAccuracy and
verticalAccuracy properties are no longer writable. This is
done to align with the other properties of the class, which
are read-only. There is no real usecase for making them
writable.

Task-number: QTBUG-89874
Change-Id: I9d14ce8c558c7b50d1e1dd8a46a599cdd4811f5b
Reviewed-by: Sona Kurazyan <sona.kurazyan@qt.io>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>