QtPositioning: update android plugin to use QJniEnvironment
This is the first patch in the series. It updates the positioning
Android plugin to use the new public QJniEnvironment class instead
of the private QtCore API.
However we can't completely remove the dependency on qjnihelpers_p.h
because we still need it to request permissions. This part will be
addressed in a separate patch.

Task-number: QTBUG-92475
Change-Id: I270c6f5866e24145d62a501f35642a67c3969b4b
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>