QDeclarativePositionSource: add property bindings
This commit ports most of the properties in QDeclarativePositionSource
to new property system.
* "active" property will be corever in the next patch.
* "updateInterval" and "preferredPositioningMethods" properties are not
converted to bindables because they have custom getters and setters,
which makes the code for porting them complicated and adds memory
overhead. However these properties are normally set once and not
updated, so there is no real benefit from converting them.

Task-number: QTBUG-89874
Change-Id: If087405c4515725ba304dd18a880cb25fbad20c0
Reviewed-by: Juha Vuolle <juha.vuolle@insta.fi>
Reviewed-by: Sona Kurazyan <sona.kurazyan@qt.io>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>