QtPositioning: fix source detection
When an 'nmeaSource' property is specified, it is read before the
component is completed. At this point, a valid position source can be
created to read the data from the specified source.
After that, when the component is completed, a new position source
is created based on the 'name' property, even if it's empty.
This results in the initial valid position source being overridden by
a default one (depending on the system configuration it may or may not
provide valid positioning information).

This patch adds a check that if nmeaSource is used, we do not try to
override it and ignore the 'name' property.

Fixes: QTBUG-85636
Change-Id: I6acdcdaf7c7f8ee168c1d48d9eb5dbc7d1838af3
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>