QGeoPositionInfoSource: fix flaky test
When introducing a dummy.source plugin for the tests, the priority was
selected the same, as for the test.source plugin. As a result, the
algorithm for selecting the default source during the test run could
pick any of them. When dummy.source was picked, the
preferredMethodsBinding test could not execute correctly, because the
dummy.source has different supported positioning methods.

This patch fixes the problem in 2 ways:
* test.source is now directly selected in the preferredMethodsBinding()
  test
* the priorities of the dummy.source and test.source plugins are
  adjusted in such a way, that test.source is picked if
  createDefaultSource() call is used.

Change-Id: Ibec456f4679207a161c2cf97c267055d75bbd03c
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>