QtPositioning tests: reduce timeouts
The tests for QGeoPositionInfoSource and QGeoSatelliteInfoSource had
the updated intervals of 7000 msec and more. According to the comments
in the code, initially it was caused by the symbian emulator, that had
a minimumUpdateInterval of 5000 msec. Currently this emulator is not
used in our CI, and the minimumUpdateInterval for all existing plugins
is not greater than 1000 msec.

This patch reduces all the intervals, which allows to reduce the total
test execution time by more than 3 times (from 930+ seconds to around
270 seconds on my local machine).

Note: this patch does not change the timeouts in the
QTRY_*_WITH_TIMEOUT calls, because they do not always stay blocked for
the whole specified timeout. Keeping large timeouts will prevent the
tests from flaky failing due to I/O operations being slow in case of
high CI load.

Fixes: QTBUG-93030
Change-Id: I0fac431448b435302dec050a9448d31e1345b287
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Juha Vuolle <juha.vuolle@insta.fi>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>