QtPositioning: support background location requests
Once running as a service, the application can only check for the
existing permissions, but not request them. Previously positioning
Android plugin was not handling this case correctly.

This patch fixes it. Now when the code is running from a service, we
only check for the required permission.

Task-number: QTBUG-71396
Pick-to: 6.2
Change-Id: Ica147a362e90624dd1a9c1c11f0ff3577991f433
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>