QtPositioning: add GnssStatus support for android plugin
This commit adds support for GnssStatus and GnssStatus.Callback API
to detect satellite info changes since API Level 24.
Note that we also need to retain the old GpsStatus API, because we
support API Level 23 as well.

This commit introduces an abstract PositioningLooperBase class on
the Java side, and creates different subclasses for each of the
APIs. Only one subclass is instantiated at each run, depending on
the OS version.
As a drive-by: removed an unused LocationManager variable during
PositioningLooper refactoring.

Fixes: QTBUG-92965
Pick-to: 6.2
Change-Id: I8af1c37d99b75fc9ad27b77b1befa4e1f8626803
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>