QtPositioning: emit missing UpdateTimeoutError during regular updates
The documentation tells that UpdateTimeoutError will be raised if the
plugin determines that it fails to provide regular updates.
This was not implemented for Android plugin so far.

This patch adds such feature. However, there is no straightforward way
to find out that the Android OS is not able to provide location updates.
For example, analyzing satellite info is not the correct approach,
because the OS also uses network information to provide the location.

As a result, the only reasonable way is to actually detect that there
were no location updates within a certain period of time.
This is the approach that is implemented in this patch. The period is
calculated based on the selected updateInterval() plus some constant.
The constant is large enough for the Android OS to start providing
updates under normal conditions.

Apart from that, the error is raised immediately when the required
location providers are disabled.

Fixes: QTBUG-36854
Change-Id: I14e8e95e721efd7fbb09d8d9a6e984f80f7249cc
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>
(cherry picked from commit d0351428c74dd9f6d9544ca8bdf442317e1976e8)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>