QtPositioning: fix provider disabled notifications
When starting the location updates, we register the same location
listener for multiple location providers (GPS & network). As a result,
the onProviderDisabled() callback is called multiple times, and we
might report too many "Closed" errors.
This patch fixes the logic by introducing an additional flag to track
that the native C++ callback was already invoked.

Drive-by: fix some code formatting.

Pick-to: 6.2
Change-Id: Ie8e93ae77eb1c9cf9044bf3a35b22c37e9f2a8aa
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>