QtPositioning: make comparison operators hidden friends
This patch changes all comparison operators in QtPositioning to
hidden friends, so that they can be found only by ADL and do not
pollute the global namespace.

Note that the comparison operators are removed from the
QGeoShape's child classes. Only the friend operators for the base
QGeoShape class itself are implemented. The ADL will manage to
correctly detect these operators, and then the virtual methods
of the private d-pointers will be called.

Task-number: QTBUG-95163
Pick-to: 6.2
Change-Id: Ice348a8a219395b0470cd0202b1fcc6673d575e2
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>