Gypsy plugin: introduce plugin parameters
Previously the plugin was taking the device name from a hardcoded path
in the GConf parameters.
This patch adds the possibility to specify the device in two ways:
* use 'gconfKey' parameter to specify the GConf key, that should be
  used to extract the device name;
* use 'deviceName' parameter to specify the device name directly.

If none of these parameters is specified, the old behavior is used,
i.e. the plugin uses the hardcoded GConf key to try to extract the
device name.

Task-number: QTBUG-74995
Pick-to: 6.2
Change-Id: I7c911630a2d367310e62764a1660171a6856e27b
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>