Move OpenWeatherMap support to separate backend class
The idea is to provide support for multiple backends in the weatherinfo
application.
This patch introduces class hierarchy for the backends and moves current
OpenWeatherMap support to a separate class, so that it can later be
replaced at runtime.

During the refactoring the number of requests is also optimized (we now
perform 2 requests instead of 3 to get the information about one city).

Currently the OpenWeatherMap backend is used directly as a property of
AppModelPrivate. More backends will be added in the following patches,
and a mechanism to switch between them will be provided.

The code for limiting the amount of the requests is currently removed.
It will be replaced by a weather data cache in the following patches.

As a drive-by: fixed a memory leak in the AppModel.

Task-number: QTBUG-60467
Pick-to: 6.2
Change-Id: I90daac9451e4c14749a1725c4a04afce0d4b7467
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>